Fix warning in plural rules where fraction is used
There are two languages, Lithuanian (lt) and Sinhala (si), which rely on the fraction part with leading zeros (f) for evaluating plural rules. At the time of evaluation, this part has to be a number and cannot be an empty string, which it currently is if the input is integer.